refactor: split OTD Livebook & analytics concerns
