--- test/lastfm_analytics_test.exs
@@ -1,4 +1,0 @@
-defmodule CodaTest do
-  use ExUnit.Case
-  doctest Coda
-end
